backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
neo4j>=5.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
pydantic>=2.0.0
